#117 refactor: split the config module into per module files
