+++ crates/hydebar-proto/src/config/system_info.rs
@@ -1,0 +1,155 @@
+//! Configuration for the system information module.
+
+use serde::Deserialize;
+
+/// Warning and alert thresholds for CPU load, in percent.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct SystemInfoCpu {
+    #[serde(default = "default_cpu_warn_threshold")]
+    pub warn_threshold:  u32,
+    #[serde(default = "default_cpu_alert_threshold")]
+    pub alert_threshold: u32
+}
+
+impl Default for SystemInfoCpu {
+    fn default() -> Self {
+        Self {
+            warn_threshold:  default_cpu_warn_threshold(),
+            alert_threshold: default_cpu_alert_threshold()
+        }
+    }
+}
+
+/// Warning and alert thresholds for memory usage, in percent.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct SystemInfoMemory {
+    #[serde(default = "default_mem_warn_threshold")]
+    pub warn_threshold:  u32,
+    #[serde(default = "default_mem_alert_threshold")]
+    pub alert_threshold: u32
+}
+
+impl Default for SystemInfoMemory {
+    fn default() -> Self {
+        Self {
+            warn_threshold:  default_mem_warn_threshold(),
+            alert_threshold: default_mem_alert_threshold()
+        }
+    }
+}
+
+/// Warning and alert thresholds for temperature, in degrees Celsius.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct SystemInfoTemperature {
+    #[serde(default = "default_temp_warn_threshold")]
+    pub warn_threshold:  i32,
+    #[serde(default = "default_temp_alert_threshold")]
+    pub alert_threshold: i32
+}
+
+impl Default for SystemInfoTemperature {
+    fn default() -> Self {
+        Self {
+            warn_threshold:  default_temp_warn_threshold(),
+            alert_threshold: default_temp_alert_threshold()
+        }
+    }
+}
+
+/// Warning and alert thresholds for disk usage, in percent.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct SystemInfoDisk {
+    #[serde(default = "default_disk_warn_threshold")]
+    pub warn_threshold:  u32,
+    #[serde(default = "default_disk_alert_threshold")]
+    pub alert_threshold: u32
+}
+
+impl Default for SystemInfoDisk {
+    fn default() -> Self {
+        Self {
+            warn_threshold:  default_disk_warn_threshold(),
+            alert_threshold: default_disk_alert_threshold()
+        }
+    }
+}
+
+/// A single readout rendered by the system information module.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub enum SystemIndicator {
+    Cpu,
+    Memory,
+    MemorySwap,
+    Temperature,
+    Disk(String),
+    IpAddress,
+    DownloadSpeed,
+    UploadSpeed
+}
+
+/// System information module behaviour.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct SystemModuleConfig {
+    #[serde(default = "default_system_indicators")]
+    pub indicators:  Vec<SystemIndicator>,
+    #[serde(default)]
+    pub cpu:         SystemInfoCpu,
+    #[serde(default)]
+    pub memory:      SystemInfoMemory,
+    #[serde(default)]
+    pub temperature: SystemInfoTemperature,
+    #[serde(default)]
+    pub disk:        SystemInfoDisk
+}
+
+impl Default for SystemModuleConfig {
+    fn default() -> Self {
+        Self {
+            indicators:  default_system_indicators(),
+            cpu:         SystemInfoCpu::default(),
+            memory:      SystemInfoMemory::default(),
+            temperature: SystemInfoTemperature::default(),
+            disk:        SystemInfoDisk::default()
+        }
+    }
+}
+
+fn default_system_indicators() -> Vec<SystemIndicator> {
+    vec![
+        SystemIndicator::Cpu,
+        SystemIndicator::Memory,
+        SystemIndicator::Temperature,
+    ]
+}
+
+fn default_cpu_warn_threshold() -> u32 {
+    60
+}
+
+fn default_cpu_alert_threshold() -> u32 {
+    80
+}
+
+fn default_mem_warn_threshold() -> u32 {
+    70
+}
+
+fn default_mem_alert_threshold() -> u32 {
+    85
+}
+
+fn default_temp_warn_threshold() -> i32 {
+    60
+}
+
+fn default_temp_alert_threshold() -> i32 {
+    80
+}
+
+fn default_disk_warn_threshold() -> u32 {
+    80
+}
+
+fn default_disk_alert_threshold() -> u32 {
+    90
+}
