+++ crates/hydebar-proto/src/config/updates.rs
@@ -1,0 +1,10 @@
+//! Configuration for the updates module.
+
+use serde::Deserialize;
+
+/// Commands used to query and apply system package updates.
+#[derive(Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct UpdatesModuleConfig {
+    pub check_cmd:  String,
+    pub update_cmd: String
+}
